Fully qualify GPOS type names in macros
Even if we define the macros lexically within the namespace curly
braces, they are still in the realm of the preprocessor and hence we
need to take extreme caution when it comes to ensuring that they are not
ambiguous.

This commit fully namespace-qualify the integral type names from GPOS to
ensure they are not ambiguous in macro expansions.